requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
alembic
psycopg2-binary
python-docx
pdfplumber
pytesseract
Pillow
python-dotenv
pydantic
requests
scikit-learn
numpy
pytest
httpx
openai
python-dateutil
gunicorn
